resources: refactored resources builder,added utility functions
